feat: enhance agent functionality and error handling
- Added new scripts `seller-sms.ts` and `buyer-agent.ts` to support SMS communication and buyer agent processes.
- Updated `package.json` to include new integration scripts for improved testing capabilities.
- Refined error handling in `seller-sms.ts` to provide clearer messages for insufficient allowance and balance.
- Enhanced `preflight.ts` to ensure type safety with bigint for bond calculations.
- Introduced a live stream feature in `agent-chat.tsx` to handle real-time events from the buyer agent, improving user interaction and feedback.

+++ quittance-agents/scripts/preflight.ts
@@ -56,8 +56,8 @@ async function main() {
   // ── Print balances ─────────────────────────────────────────────────────────
   const [buyerKite,  buyerPyusd]  = await Promise.all([provider.getBalance(buyer.address),  c.pyusd.balanceOf(buyer.address)]);
   const [sellerKite, sellerPyusd] = await Promise.all([provider.getBalance(seller.address), c.pyusd.balanceOf(seller.address)]);
-  const sellerBond = await c.bond.bonds(seller.address);
-  const minBond    = await c.bond.MIN_BOND();
+  const sellerBond = (await c.bond.bonds(seller.address)) as bigint;
+  const minBond    = (await c.bond.MIN_BOND()) as bigint;
 
   console.log("── Balances ──────────────────────────────────────");
   console.log(`Buyer  (${buyer.address})`);
